Added unpolished version of 2D MCMC optimization
This is simply a toy model as a proof of concept. To be adapted to be used on camera-SNR optimization in a 3D parameter space. This is currently unpolished, there are a few more functions I could create to streamline the calculations and final output. NOTE: the animations would not be required for the 3D camera parameter space optimization as it would be hard to visualize.